analyzer/requirements.txt
# Innervision Synth Endpoint - Analyzer Dependencies
# Video analysis and workflow extraction pipeline

# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
click>=8.1.7
httpx>=0.27.0  # Async HTTP client for VLM APIs

# Video Processing
opencv-python>=4.9.0
ffmpeg-python>=0.2.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Image Analysis & ML
numpy>=1.26.0
pillow>=10.2.0
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0  # YOLO for UI detection
transformers>=4.37.0
timm>=0.9.12

# OCR
pytesseract>=0.3.10
easyocr>=1.7.1

# Data Processing
pandas>=2.2.0
scikit-learn>=1.4.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Development
black>=24.1.0
ruff>=0.2.0
mypy>=1.8.0
